refactor: Service 테스트를 위한 Dao 추상화 및 FakeTimeDao 구현
